requirements.txt
python-telegram-bot>=20.7
python-dotenv>=1.1.0
Pillow>=10.0.0
